Create 15%_CST
A CST (CST stands for "Central Sales Tax") from your grocery shopping of 15% (fake law used), of the total amount of balance. If the CST is larger than the balance or the CST percentage is higher than 15%, the program will print one line of code for each of the above mentioned invalid inputs. Other than that it'll work fine.